Fixed Lexer Bugs, changed tests

+++ src/main/java/com/AdiAyush/EZPZ/Lexer/TokenType.java
@@ -22,7 +22,7 @@ public enum TokenType {
         STRING,
         BOOL,
         CHAR,
-        DECIMAL,
+        FLOAT,
         LONG,
         VOID,
         CONSTANT,
